Add Formatter io::Write decorator
